requirements.txt
streamlit
langchain
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
PyPDF2
python-docx
openpyxl

langchain-openai
